feat(images): Centralize ImageLoader and remove feature-specific loaders
This commit centralizes the `ImageLoader` creation and management into the `app` module. Feature modules (`article` and `articles`) no longer define their own `ImageLoader` instances or related dependencies like `OkHttpClient` and `Mapper`. Instead, they will consume the `ImageLoader` provided by the `app` module via a `CompositionLocal`.

- Added `ImageLoaderModule` to the `app` module to provide a singleton `ImageLoader`, `OkHttpClient`, and `Mapper`.
  - The `OkHttpClient` is configured with an `HttpLoggingInterceptor`.
- Added `coil-core`, `coil-network-okhttp`, and `okhttp-logging` dependencies to the `app` module.
- Added `LocalImageLoader` `CompositionLocal` to `core:uisystem` to provide the `ImageLoader` instance down the Compose tree.
- Updated `MainActivity` to provide the `ImageLoader` using `CompositionLocalProvider`.
- Removed `ImageLoader`, `OkHttpClient`, and `Mapper` provider methods from `ArticleModule` and `ArticlesModule`.
- Removed `ArticleEntryPointImageLoader` and `ArticlesEntryPointImageLoader` qualifiers and their usage in `ArticleEntryPoint` and `ArticlesEntryPoint`.
- Removed `coil-core` and `coil-network-okhttp` dependencies from `feature:article` and `feature:articles` modules.
- Updated `ArticleScreen` and `ArticlesScreen` to use the `LocalImageLoader` from `core:uisystem` instead of feature-specific local providers.
- Updated `Article` component in `feature:articles` to use the centralized `LocalImageLoader`.
- Added `coil-core` dependency to `core:uisystem`.
