+++ core/uisystem/src/main/kotlin/com/kesicollection/core/uisystem/LocalImageLoader.kt
@@ -1,0 +1,8 @@
+package com.kesicollection.core.uisystem
+
+import androidx.compose.runtime.compositionLocalOf
+import coil3.ImageLoader
+
+val LocalImageLoader = compositionLocalOf<ImageLoader> {
+    error("No ImageLoader provided")
+}
