+++ feature/article/build.gradle.kts
@@ -9,8 +9,6 @@ android {
 
 dependencies {
     implementation(libs.coil.compose)
-    implementation(libs.coil.core)
-    implementation(libs.coil.network.okhttp)
     implementation(libs.markdown)
     testImplementation(libs.coil.test)
 }
