+++ feature/articles/build.gradle.kts
@@ -9,7 +9,5 @@ android {
 
 dependencies {
     implementation(libs.coil.compose)
-    implementation(libs.coil.core)
-    implementation(libs.coil.network.okhttp)
     testImplementation(libs.coil.test)
 }
